+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]## [0.5.7.2]
+[Unreleased]## [0.5.8.1b0]
+
+## [0.5.8.0]- 2026-01-23
+### Fixed
+- Sphinx HTML documentation (due to incompatibilities between pandas and numpy versions)Spi
+
+## [0.5.8.0b0]- 2026-01-16
+### Fixed
+- ScenarioManager migrate project_lib to ibm-watson-studio-lib for CPD-SaaS compatibility
+- Removed deprecated ScenarioManager.add_data_file_to_project_s()
+- Deprecated and removed support for CPD25 in ScenarioManager
+- ScenarioManager.load_data_from_excel in CPDaaS loads Excel file using ibm-watson-studio-lib
+- ScenarioManager.write_data_to_excel in CPDaaS saves Excel file using ibm-watson-studio-lib (using add_file_as_data_asset)
+- ScenarioManager.get_dd_client migrated to ibm-watson-studio-lib
 
 ## [0.5.7.2]- 2026-01-13
 - Released version 0.5.7.2b2 as 0.5.7.2